Test: Reset project structure to multiple modules
Since the structure didn't seem to solve the problem causing the pipeline to fail the build, I reverted it back to its original state.

As the application modules require a main class, I implemented a class which just prints a text
message as the run task is started.

To actually start the modules using gradle I had to set the main class in the build.gradle.