In this commit-message the description of the change before is also included, as I forgot to place it into the message box.
javafx included in build.gradle of gui module

editor window designed in fxml-editor.

introduction of the class elementsInPlayfield that defines the content all panes in the gridPane.

introduction of the enum elementInPlayfield that all allows to distinguish between all different graphic elements in one pane in the gridPane.

introduction of the static class storeLastKey that allows to memorize the last pressed key before any other event.
This is used in the mousePressedEvent to read the character that was pressed before.
This is to distinguish between the different elementInPlayfield that the user selects by keyboard.

Introduction of two static HashMaps in the class MethodsForElements:
The HashMap elementInPlayfieldBackgroundFillMap maps the enum ElementInPlayfield to the BackgoundFill. backgroundFill ist used when a background is generated. The background is needed to set the background to the pane.

The HashMap STRING_ELEMENT_IN_PLAYFIELD_MAP maps the of the keyCode to an elementInPlayfield. This allows the user to select and deselect an element

Intoduction of the controler class EditorWindowControler:
The initialialize method calls set panesOnPlayfield.
In this method all for each element in the gridPane a pane is placed.
Each plane displays the graphic that is assoziated this the elementOfPlayfield at this place.
Together with placing the pane in the gridPane the OnMousedPessedEvent for the pane where the mouse is, will call the method setResetElement.

If the mouse is pressed the first time the element is set to the element assoziated with the letter pressed before.  (not jet implemented)
It the mouse is pressed a second time the element is set to NoElement of the enums elementInPlayfield.

The class Floor will be replace by elementsOfPlayfield.  (not jet implemented)