This commit adds the playfield Window. It is opened by the Play Button in the Editor windows. This calls the method startPayfieldWindow(). For the OnKeyPressed Event an event handler is started and the window PlayfieldWindow is opened. In initialize method of the PlayfieldWindowControler all panes are are set up and are initialsized according to the background defined by elementsInPlayfield.
The controler also contains the static public method changeAllPanesDependingOnElementsInPlayfied()
This method can be called for extern to update the window when the content of elementsInPlayfield is changed.

This commit also contains a DummyGameForTesting
In the event handler the method moveOwl(KeyCode keyCode) is set.
If the keys 8,4,6,2 are pressed the owl is moved in the elementsInPlayfield Array.
Then the changeAllPanesDependingOnElementsInPlayfied() is called and the window is updated.