Refactoring and introduced special LevelJSON class
The Level object itself holds information which does not need to be
saved or transmitted through the network.
The LevelJSON class only stores the actually used data of the level
layout and provides a factory method from itself to a Level object.

TODO: Change Level constructor so it produces a default blank level
and introduce a demo level factory method.