Edited server to use Connection class and react to messages
Because a Connection class for handling the streams between client
and server already exists, this one is now used by the server.

Also, a connection manager class got introduced to handle multiple
clients requesting on the same server.

The server currently only reacts to a connection request with a
notification and kills the connection (as the client on the other
side stops after sending his connection request).