Enable saving from client to server, bugfixes
Annoying bug where Server couldn't load freshly created levels due to
the file not being present for the server. The fix was rather simple:
As there already was a map from the level name to the level savefile,
the load request got rerouted to the stored savefile rather than
reading from the actual storage.

The persistent saves will get updated to the local save of the server
by boot up - so at all times all level savefiles are present and
persistent (as far as testing goes).